Uses/electronfix.mk: Framework for quick porting of Electron apps.
Summary:
The idea behind this USES is to take a binary distribution of some Electron app
and replace Linux binaries with the native ones. Not only this allows for
quick porting, but it is also a way to get closed-source Electron apps (Obsidian,
XMind) working on FreeBSD.

The deskutils/obsidian update and the graphics/drawio addition gives an example
on how to use electronfix.

The stage-qa spits a lot of warnings about missing dependencies, but all of these
are false positives coming from copying the Electron main binary. Since the
framework makes the port RUN_DEPENDS on the Electron package, these dependencies
are guaranteed to be fulfilled anyways.

Disregard the logseq addition for now, as it is not 100% ready.

Reviewers: #contributor_reviewers_ports, tagattie

Differential Revision: https://reviews.freebsd.org/D42408